requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
django==4.2.6
    # via
    #   -r requirements.in
    #   django-import-export
django-environ==0.11.2
    # via -r requirements.in
django-import-export==3.3.8
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==21.2.0
    # via -r requirements.in
markuppy==1.14
    # via tablib
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==23.2
    # via gunicorn
pillow==10.3.0
    # via -r requirements.in
psycopg==3.1.12
    # via -r requirements.in
pyyaml==6.0.1
    # via tablib
sqlparse==0.4.4
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via django-import-export
typing-extensions==4.8.0
    # via
    #   asgiref
    #   psycopg
whitenoise==6.6.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
